Check if eform nginx config does exist

+++ install-eforms.sh
@@ -35,15 +35,21 @@ read -e -p "Please enter the public host name for Camunda server (fully qualifie
 read -e -p "Please enter the public host name for Alfresco server (fully qualified domain name)${ques} [`hostname`] " -i "`hostname`" ALFRESCO_HOSTNAME
 	
 if [ -f "/etc/nginx/sites-available/$CAMUNDA_HOSTNAME.conf" ]; then
-	#sudo sed -i "0,/server/s/server/upstream eform {	    \n\tserver localhost:$TOMCAT_HTTP_PORT;	\n}\n\n&/" /etc/nginx/sites-available/$CAMUNDA_HOSTNAME.conf
-	sudo sed -i "1 i\upstream eform {	    \n\tserver localhost:$TOMCAT_HTTP_PORT;	\n}\n\n" /etc/nginx/sites-available/$CAMUNDA_HOSTNAME.conf
- 
- # Insert camunda configuration content before the last line in domain.conf in nginx
- #sudo sed -i "$e cat $NGINX_CONF/sites-available/camunda.conf" /etc/nginx/sites-available/$hostname.conf
- sudo mkdir temp
- sudo cp $NGINX_CONF/sites-available/eform.snippet	temp/
- sudo sed -e '/##EFORM##/ {' -e 'r temp/eform.snippet' -e 'd' -e '}' -i /etc/nginx/sites-available/$CAMUNDA_HOSTNAME.conf
- sudo rm -rf temp
+
+	# Check if eform config does exist
+	eform_found=$(grep -o "eform" $CATALINA_HOME/conf/server.xml | wc -l)
+	
+	if [ $camunda_found = 0 ]; then
+		#sudo sed -i "0,/server/s/server/upstream eform {	    \n\tserver localhost:$TOMCAT_HTTP_PORT;	\n}\n\n&/" /etc/nginx/sites-available/$CAMUNDA_HOSTNAME.conf
+		sudo sed -i "1 i\upstream eform {	    \n\tserver localhost:$TOMCAT_HTTP_PORT;	\n}\n\n" /etc/nginx/sites-available/$CAMUNDA_HOSTNAME.conf
+		 
+		 # Insert camunda configuration content before the last line in domain.conf in nginx
+		 #sudo sed -i "$e cat $NGINX_CONF/sites-available/camunda.conf" /etc/nginx/sites-available/$hostname.conf
+		 sudo mkdir temp
+		 sudo cp $NGINX_CONF/sites-available/eform.snippet	temp/
+		 sudo sed -e '/##EFORM##/ {' -e 'r temp/eform.snippet' -e 'd' -e '}' -i /etc/nginx/sites-available/$CAMUNDA_HOSTNAME.conf
+		 sudo rm -rf temp
+	fi
 fi
 
 echo "We are waiting for eform being deployed...."
@@ -56,3 +62,5 @@ sudo sed -i "s/\(^CmisRepoId=\).*/\1-default-/" 	$CATALINA_HOME/webapps/eform/WE
 sudo sed -i "s/\(^CmisUser=\).*/\1admin/"  	$CATALINA_HOME/webapps/eform/WEB-INF/classes/application.properties
 sudo sed -i "s/\(^CmisPassword=\).*/\1admin/"  	$CATALINA_HOME/webapps/eform/WEB-INF/classes/application.properties
 sudo sed -i "s/\(^CmisRootFolder=\).*/\Data Dictionary/"  $CATALINA_HOME/webapps/eform/WEB-INF/classes/application.properties
+
+. $DEVOPS_HOME/devops-service.sh restart
